third_party/blink/web_tests/external/wpt/html/infrastructure/urls/resolving-urls/query-encoding/windows-1251_include=xhr-expected.txt

This is a testharness.js-based test.
FAIL XMLHttpRequest#open() assert_equals: expected "%C3%A5" but got "%26%23229%3B"
Harness: the test ran to completion.
